requirements.txt
# PyPDF2
# Pillow
# img2pdf
# reportlab
# tk

flask==2.3.2
PyPDF2==3.0.1
Pillow==9.5.0
reportlab==3.6.12
img2pdf==0.4.4
werkzeug==2.3.4